Review suggestions & fixes
