+++ src/bin/package_for_release/bundle_macos.rs
@@ -13,20 +13,20 @@ const INTEL_MAC_TARGET: &str = "x86_64-apple-darwin";
 
 pub fn bundle_macos_app(enable_web_rendering: bool) -> Result<()> {
     let output_name = match enable_web_rendering {
-        true => root_dir.join("video_compositor_with_web_rendering"),
-        false => root_dir.join("video_compositor"),
+        true => "video_compositor_with_web_rendering",
+        false => "video_compositor",
     };
 
     if cfg!(target_arch = "x86_64") {
         bundle_app(
             INTEL_MAC_TARGET,
-            format!("{output_name}_x86_64.tar.gz"),
+            &format!("{output_name}_darwin_x86_64.tar.gz"),
             enable_web_rendering,
         )?;
     } else if cfg!(target_arch = "aarch64") {
         bundle_app(
             ARM_MAC_TARGET,
-            format!("{output_name}_aarch64.tar.gz"),
+            &format!("{output_name}_darwin_aarch64.tar.gz"),
             enable_web_rendering,
         )?;
     } else {
@@ -35,11 +35,7 @@ pub fn bundle_macos_app(enable_web_rendering: bool) -> Result<()> {
     Ok(())
 }
 
-fn bundle_app(
-    target: &'static str,
-    output_name: &'static str,
-    enable_web_rendering: bool,
-) -> Result<()> {
+fn bundle_app(target: &'static str, output_name: &str, enable_web_rendering: bool) -> Result<()> {
     let root_dir_str = env!("CARGO_MANIFEST_DIR");
     let root_dir: PathBuf = root_dir_str.into();
     let build_dir = root_dir.join(format!("target/{target}/release"));
@@ -53,8 +49,8 @@ fn bundle_app(
         use compositor_chromium::cef;
 
         info!("Build process_helper binary.");
-        utils::cargo_build("process_helper", target)?;
-        cef::bundle_app(&build_dir, &tmp_dir.join("video_compositor.app"), false)?;
+        utils::cargo_build("process_helper", target, false)?;
+        cef::bundle_app(&build_dir, &tmp_dir.join("video_compositor.app"))?;
     } else {
         let _ = fs::remove_dir_all(&tmp_dir);
         fs::create_dir_all(&tmp_dir)?;
